+++ Assets/_BForBoss/_Core/Scripts/Managers/RaceCourseWorldManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30f55d712629740c8a40b60acf6366d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
